requirements.txt
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.2.1
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
xgboost==2.0.2
